refactor: hierachy
